Review fixes: enforceable whole cpus, Harbor-config assertions, asset-instruction regression test
- ExecutionPolicy.cpus is now an integer: Harbor's
  EnvironmentConfig.override_cpus is an int, so a fractional declaration
  could never be enforced (declared-is-enforced, #36 reopen)
- the e2e pi profile declares resources {cpus: 1, memory_mb: 256} so the
  full chain exercises the override application against real Docker; the
  full-chain test asserts the timeout lands in the generated task.toml
  and instruction.md carries exactly the pinned prompt bytes
- new e2e regression test: an asset-resolved instruction (#20 reopen)
  verifies against the profile's prompt digest — matching bytes run,
  wrong digest fails pre-agent with environment_invalid and no model call

+++ src/aco/models.py
@@ -47,7 +47,9 @@ class ExecutionPolicy(BaseModel):
 
     model_config = ConfigDict(extra="forbid")
 
-    cpus: float | None = Field(default=None, gt=0, le=64)
+    # cpus is an integer: Harbor's EnvironmentConfig.override_cpus is an int,
+    # so a fractional declaration could never be enforced (#36 reopen)
+    cpus: int | None = Field(default=None, gt=0, le=64)
     memory_mb: int | None = Field(default=None, gt=0, le=65536)
     timeout_sec: int | None = Field(default=None, gt=0, le=86400)
     network: Literal["offline", "online"] | None = None
